requirements.txt
streamlit
openai
python-dotenv
pymupdf
pytesseract
Pillow
openpyxl
